Completed 01-css-methods
I created a new file called style.css and made changes to the div
element and changed the background color to red. I also changed the
font color to white and the font size to 32px.

For the second part of the assignment I added in a internal CSS to
the index.html file and made the same changes to the p element.
I changed the background color to green, the font color to white, and
and the font size to 18px.

For the third part of the assignment I added in an inline CSS to the
button element and made changed the background to orange and the font
size to 18px.