Modified the index.html and style.css file
For odd numbered elements, I added a class called odd and set the
background color to #FFB6C1, added a list of fonts containing 'Verdana'
and 'DejaVu Sans', with 'sans-serif' as the fallback.

For the second element, I added a class called two and set the text to
blue and the font size to 36px.

For the third element, I added an id called additional and made the font
size 24px.

For the fourth and final element, I added a class called four and set
the background to light green, the font size to 24px and the font weight
to bold.