Change the color of the buttons
For both elements, I was able to change the font size to 28px and the
font familyt to Helvetica and Time New Roman, with san-serif as the
fallback.

For the first element, I changed the background to black and the text to
white.

For the second element, I changed the background to yellow and end
ensured the text was black.